+++ src/main/java/com/telepathicgrunt/blame/main/BiomeBlame.java
@@ -3,6 +3,8 @@ package com.telepathicgrunt.blame.main;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonElement;
+import com.mojang.datafixers.util.Either;
+import com.mojang.serialization.DataResult;
 import com.mojang.serialization.JsonOps;
 import com.telepathicgrunt.blame.Blame;
 import net.minecraft.crash.CrashReport;
@@ -16,9 +18,6 @@ import net.minecraft.world.gen.feature.ConfiguredFeature;
 import net.minecraft.world.gen.feature.structure.Structure;
 import net.minecraftforge.registries.ForgeRegistries;
 import org.apache.logging.log4j.Level;
-
-import java.util.Arrays;
-import java.util.Optional;
 
 /* @author - TelepathicGrunt
  *
@@ -35,50 +34,33 @@ public class BiomeBlame {
      * Prints registry name of feature and biome.
      * Prints the crashlog to latest.log as well.
      */
-    public static void addFeatureDetails(Biome biome, WorldGenRegion worldGenRegion, ConfiguredFeature<?, ?> configuredFeature, CrashReport crashreport) {
+    public static void addFeatureDetails(Biome biome, WorldGenRegion worldGenRegion, ConfiguredFeature<?, ?> configuredFeature) {
         DynamicRegistries dynamicRegistries = worldGenRegion.registryAccess();
         Gson gson = new GsonBuilder().setPrettyPrinting().create();
-        ResourceLocation configuredFeatureID = null;
-        ResourceLocation biomeID = null;
-        Optional<JsonElement> configuredFeatureJSON = Optional.empty();
-        StackTraceElement[] trace = null;
+        String report;
 
-        try {
-            configuredFeatureID = dynamicRegistries.registryOrThrow(Registry.CONFIGURED_FEATURE_REGISTRY).getKey(configuredFeature);
-            if (configuredFeatureID == null) {
-                configuredFeatureID = WorldGenRegistries.CONFIGURED_FEATURE.getKey(configuredFeature);
-            }
-            biomeID = dynamicRegistries.registryOrThrow(Registry.BIOME_REGISTRY).getKey(biome);
-        }
-        catch (Throwable ignored) {
-        }
+        ResourceLocation configuredFeatureID = dynamicRegistries.registryOrThrow(Registry.CONFIGURED_FEATURE_REGISTRY).getKey(configuredFeature);
+        if (configuredFeatureID == null) configuredFeatureID = WorldGenRegistries.CONFIGURED_FEATURE.getKey(configuredFeature);
 
-        try {
-            configuredFeatureJSON = ConfiguredFeature.CODEC.encode(() -> configuredFeature, JsonOps.INSTANCE, JsonOps.INSTANCE.empty()).get().left();
-        }
-        catch (Throwable e) {
-            trace = e.getStackTrace();
-        }
+        ResourceLocation biomeID = dynamicRegistries.registryOrThrow(Registry.BIOME_REGISTRY).getKey(biome);
+        Either<JsonElement, DataResult.PartialResult<JsonElement>> configuredFeatureJSON = ConfiguredFeature.CODEC.encode(() -> configuredFeature, JsonOps.INSTANCE, JsonOps.INSTANCE.empty()).promotePartial(s -> {}).get();
 
         // Add extra info to the crash report file.
         if (configuredFeatureID == null) {
-            crashreport.getSystemDetails()
-                    .setDetail("\n****************** Blame Report ConfiguredFeature " + Blame.VERSION + " ******************",
-                            "\n\n ConfiguredFeature name was unable to be found due to either the configuredfeature registry or " +
-                                    "\n biome registry missing somehow. Or that the configuredfeature is not in any registries." +
-                                    "\n Sorry but Blame isn't really able to get much info but..." +
-                                    "\n Here's the best attempt at turning the configuredfeature to JSON for analysis: \n" + (configuredFeatureJSON.isPresent() ? gson.toJson(configuredFeatureJSON.get()) : ""));
+            report = "\n****************** Blame Report ConfiguredFeature " + Blame.VERSION + " ******************" +
+                    "\n\n ConfiguredFeature name was unable to be found due to the configuredfeature not being registered." +
+                    "\n Sorry but Blame isn't really able to get much info but..." +
+                    "\n Here's the best attempt at turning the ConfiguredFeature to JSON for analysis: \n" + gson.toJson(configuredFeatureJSON.left().get());
         }
         else {
-            crashreport.getSystemDetails()
-                    .setDetail("\n****************** Blame Report ConfiguredFeature " + Blame.VERSION + " ******************",
-                            "\n\n ConfiguredFeature Registry Name : " + configuredFeatureID +
-                                    "\n Biome Registry Name : " + (biomeID != null ? biomeID : "Wait what? How is the biome not registered and has no registry name!?!? This should be impossible!!!") +
-                                    "\n\n JSON info : " + (configuredFeatureJSON.isPresent() ? gson.toJson(configuredFeatureJSON.get()) : "Failed to get JSON somehow. Stacktrace of error:\n" + Arrays.toString(trace)) + "\n\n");
+            report = "\n****************** Blame Report ConfiguredFeature " + Blame.VERSION + " ******************" +
+                    "\n\n ConfiguredFeature Registry Name : " + configuredFeatureID +
+                    "\n Biome Registry Name : " + (biomeID != null ? biomeID : "Wait what? How is the biome not registered and has no registry name!?!? This should be impossible!!!") +
+                    "\n\n JSON info : " + (configuredFeatureJSON.left().isPresent() ? gson.toJson(configuredFeatureJSON.left().get()) : "Failed to get JSON somehow. Error:\n" + configuredFeatureJSON.right().toString()) + "\n\n";
         }
 
         // Log it to the latest.log file as well.
-        Blame.LOGGER.log(Level.ERROR, crashreport.getFriendlyReport());
+        Blame.LOGGER.log(Level.ERROR, report);
     }
 
 
@@ -90,16 +72,8 @@ public class BiomeBlame {
     public static void addStructureDetails(Biome biome, WorldGenRegion worldGenRegion, Structure<?> structureFeature, CrashReport crashreport) {
         DynamicRegistries dynamicRegistries = worldGenRegion.registryAccess();
 
-        ResourceLocation structureID = null;
-        ResourceLocation biomeID = null;
-
-        try {
-            structureID = ForgeRegistries.STRUCTURE_FEATURES.getKey(structureFeature);
-            biomeID = dynamicRegistries.registryOrThrow(Registry.BIOME_REGISTRY).getKey(biome);
-        }
-        catch (Throwable ignored) {
-        }
-
+        ResourceLocation structureID = ForgeRegistries.STRUCTURE_FEATURES.getKey(structureFeature);
+        ResourceLocation biomeID = dynamicRegistries.registryOrThrow(Registry.BIOME_REGISTRY).getKey(biome);
 
         // Add extra info to the crash report file.
         // Note, only structures can do the details part as configuredfeatures always says the ConfiguredFeature class.
@@ -107,7 +81,7 @@ public class BiomeBlame {
                 .setDetail("\n****************** Blame Report ConfiguredStructure " + Blame.VERSION + " ******************",
                         "\n\n Structure Name : " + structureFeature.getFeatureName() + // Never null
                                 "\n Structure Registry Name : " + (structureID != null ? structureID : "Structure is not registered somehow. Yell at the mod author when found to register their structures!") +
-                                "\n Structure Details : " + structureFeature.toString() +
+                                "\n Structure Details : " + structureFeature +
                                 "\n Biome Registry Name : " + (biomeID != null ? biomeID : "Wait what? How is the biome not registered and has no registry name!?!? This should be impossible!!!"));
 
         // Log it to the latest.log file as well.